pyproducer/pyproject.toml
[tool.poetry]
name = "pyproducer"
version = "0.1.0"
description = ""
authors = ["tgeremias <tgeremias@grupodimed.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
